requirements.txt
amqp==5.3.1
anyio==4.9.0
APScheduler==3.11.2
asgiref==3.8.1
attrs==25.3.0
autopep8==2.3.2
beautifulsoup4==4.13.4
billiard==4.2.1
black==24.3.0
bleach==6.3.0
celery==5.4.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.5
django-ckeditor==6.7.3
django-ckeditor-5==0.2.13
django-ckeditor5==0.0.4
django-cors-headers==4.3.1
django-jazzmin==2.6.1
django-js-asset==3.1.2
django-modeltranslation==0.18.11
django-redis==5.4.0
django-rosetta==0.10.0
django-silk==5.3.1
django-summernote==0.8.20.0
django-unfold==0.46.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
drf-spectacular-sidecar==2024.3.4
flower==2.0.1
gprof2dot==2025.4.14
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.12.3
idna==3.10
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.3.8
polib==1.2.0
prometheus_client==0.22.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.9
pycodestyle==2.13.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==22.1
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
tornado==6.5.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
